+++ thermostate/__init__.py
@@ -1,1 +1,2 @@
 from .thermostate import units, Q_, State  # noqa
+from .units import EnglishEngineering, SystemInternational  # noqa