+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- Common unit abbreviations in thermostate.EnglishEngineering and thermostate.SystemInternational
 
 ### Changed